moved color mode dropdown to settings
